+++ database/migrations/2015_12_29_091221_add_agreements_to_relation_table.php
@@ -14,7 +14,7 @@ class AddAgreementsToRelationTable extends Migration
     {
         Schema::table('relations', function($table)
         {
-            $table->integer('agreement_id')->unsigned();
+            $table->bigInteger('agreement_id')->unsigned();
             $table->foreign('agreement_id')->references('id')->on('agreements')->onDelete('cascade');
         });
     }
